Removes CreateSrcFiles and CreateNewSrcFiles
CSrcFiles already handles filling in an empty Files: list, so no reason to duplicate that -- so I removed CreateNewSrcFiles().

CreateSrcFiles() was a hack to allow for reconverting a .srcfiles file from whatever script it was originally converted from. While that is potentially useful, it needs a LOT more work to be useful for anything but testing.

The rest of the changes are a first pass at cleanuing up variable names to match what they are actually being used for.