Overhauls the code for producing tasks (.vscode/tasks.json)
The makefile can now be anywhere.

On Windows, we use nmake for the MSVC compiler — since the makefile sets the compiler to msvc because nmake is running it.

On Windows, if clang and mingw32-make are available, we generate targets using clang for debug, release, and rebuild. If clang is not available, but gcc is, then we add targets for the GCC compiler instead.

For non-Windows build, we generate targets as if they were using nmake and MSVC, and then fix them up just before writing to use make.exe nad either clang or gcc.