Changes the way we look for .srcfiles.yaml
Instead of passing a flag that determines whether to start looking in .vscode, we instead pass an option ttCStr pointer that points to a directory to use as the root for a search. It's now valid to specify ".private" as the root, and have the actual .srcfiles.yaml be in ".private/.vscode".