Adds environment variable flags to compiler line
Checks CFLAGS, CFLAGSR and CFLAGSD and adds them to the .ninja file and .vscode/c_cpp_properties.json as if they had been specified in the Options: section of .srcfiles.yaml.

We don't track who the caller is, so these will get processed from both types of makefile. If that's a problem, will need to raise a separate issue and try to track the type of caller.

Fixes # 129