Create msvcBuild64D.ninja
Normally we wouldn't track a .ninja script until we are ready to go public. However, if in the process of testing we create invalid .ninja scripts and we don't have a backup makeninja.exe lying around to fix the problem, then we need some way to restore our ability to build at least makeninjad.exe.