Stop using wxWidgets on Windows builds
With C++17 providing file system functions, and ttlib::dlg handling UI, there
isn't anything from wxWidgets that we need.

Some code is still conditionalized for now because we will need wxWidgets
to handle the UI on non-Windows platforms. It won't be using this code
base, but it will need to be copied, so we'll leave the conditionalized code
for now.