Fix handling of c_cpp_properties.json
New version now uses ${workspaceRoot}/ as the root for includes instead
of the absolute path. It also defaults to c++11 instead of c++17 for the
cppStandard, and changes that if std:c++ was specified as a common
flag.

The updater correctly locates definitions in the various Cflags settings,
and makes includes relative to the location of the project file.