MS_Linker should default to true
For several versions the LLVM linker has failed on projects that use
wxWidgets because of the manifest directive. Even when it finally gets
fixed, we can't know if everyone using the project file will have a version
that supports it. The safer option is to default to using the MS linker when
compiling with clang-cl